gateway/requirements.txt
# ECOBIN Gateway — Dependências Python
# ======================================

# MQTT Client (v2.0+ com CallbackAPIVersion)
paho-mqtt>=2.0.0

# Google Gemini AI (SDK moderno)
google-genai>=1.0.0

# Processamento de imagem
Pillow>=10.0.0

# HTTP requests (captura de imagem do ESP32-CAM)
requests>=2.31.0

# Variáveis de ambiente
python-dotenv>=1.0.0

# Web Dashboard (para fase posterior)
flask>=3.0.0
flask-socketio>=5.3.0
